+++ bitblas/ops/general_matmul/tilelang/dense/matmul_wmma.py
@@ -1,0 +1,104 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+from bitblas import tvm as tvm
+
+from bitblas.base.roller.hint import Hint
+from bitblas.base.roller.rasterization import NoRasterization
+from dataclasses import dataclass
+from bitblas.tl.base_hint import BaseTLHint
+from .matmul_tensorcore import MatmulBaseScheduler
+
+# GPU warp configuration for NVIDIA GPUs
+warp_size = 32
+
+# TODO(lei): This is not implemented in the current version of the codebase
+@dataclass
+class MatmulFineGrainScheduler(MatmulBaseScheduler):
+    # Fine-grained matrix multiplication scheduler
+    # Allows for more detailed configuration.
+
+    # Tensor Core Warp Configuration
+    block_row_warps: int = 2
+    block_col_warps: int = 2
+    warp_row_tiles: int = 32
+    warp_col_tiles: int = 32
+    chunk: int = 32  # Usually determines the K-dimension split size
+
+    # Other Optimization Parameters
+    num_stages: int = 2
+    enable_rasterization: bool = False
+
+    class TLHint(BaseTLHint):
+
+        def __init__(self):
+            super().__init__()
+
+        @classmethod
+        def from_roller_hint(cls, hint: Hint):
+            tl_hint = cls()
+            for key, value in hint.__dict__.items():
+                setattr(tl_hint, key, value)
+
+            block = hint.block
+            warp = hint.warp
+            rstep = hint.rstep
+            num_stages = hint.pipeline_stage
+            rasterization_plan = hint.rasterization_plan
+            enable_rasterization = not isinstance(
+                rasterization_plan, NoRasterization
+            )
+
+            block_row_warps = block[0] // warp[0]
+            block_col_warps = block[1] // warp[1]
+            warp_row_tiles = warp[0]
+            warp_col_tiles = warp[1]
+            chunk = rstep[0]
+
+            if num_stages == 1:
+                num_stages = 0  # disable pipelining
+
+            tl_hint.block_row_warps = block_row_warps
+            tl_hint.block_col_warps = block_col_warps
+            tl_hint.warp_row_tiles = warp_row_tiles
+            tl_hint.warp_col_tiles = warp_col_tiles
+            tl_hint.chunk = chunk
+            tl_hint.num_stages = num_stages
+            tl_hint.enable_rasterization = enable_rasterization
+
+            return tl_hint
+
+        def get_config_params(self):
+            return {
+                "block_row_warps": self.block_row_warps,
+                "block_col_warps": self.block_col_warps,
+                "warp_row_tiles": self.warp_row_tiles,
+                "warp_col_tiles": self.warp_col_tiles,
+                "chunk": self.chunk,
+                "num_stages": self.num_stages,
+                "enable_rasterization": self.enable_rasterization,
+            }
+
+        def __repr__(self):
+            return (
+                "{"
+                f"block_M={self.block_row_warps * self.warp_row_tiles},"
+                f"block_N={self.block_col_warps * self.warp_col_tiles},"
+                f"warp_M={self.warp_row_tiles},"
+                f"warp_N={self.warp_col_tiles},"
+                f"block_K={self.chunk},"
+                f"threads={self.block_row_warps * self.block_col_warps * warp_size},"
+                f"num_stages={self.num_stages},"
+                f"enable_rasterization={self.enable_rasterization}"
+                "}"
+            )
+
+    def __post_init__(self):
+        # Validate the matrix transpose settings
+        assert (
+            self.trans_A is False
+        ), "Currently only support Matrix A not transposed"
+        assert (
+            self.trans_B is True
+        ), "Currently only support Matrix B transposed"
+
+        return
